dockerfile: Do not rely on script being executable
The preloader container is built on the user's system. On
Windows, the files don't retain their executable flags,
thus the preloader Python script can't be started just
by running the script. Now explicitly running the script
with `python3`, so whether or not the script is executable,
it will run.

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@resin.io>