preload.py: Flush output buffer when data's ready
When attaching to a docker container, docker will buffer
the standard input/output. On Windows the script seemed to hang
as the reply from the preloader container might be quite small,
thus docker just keeps waiting for more data before giving it
to the program that attached to it. Flushing stdout seems to
have the desired effect and the caller receives the reply
from the preloader.

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@resin.io>